requirements.txt
chainlit==0.2.109
langchain==0.0.189
openai==0.27.6
python-dotenv==1.0.0
